Handle `Actions` from a sink synchronously.
This guarantees that an action being applied from a sink
will immediately cause a render pass, providing a new sink to
the UI side, and prevents lost events.

Additionally, fixes that the SignalProducer from a worker
was not being cancelled on a subsequent render pass if it
was not declared.

Fixes #348